Prepare all rights
