+++ config.php.local
@@ -11,12 +11,24 @@ define("DB_CHARSET","utf8");
 define("NUMBER_ARTICLE_PER_PAGE",5);
 
 // Upload images path
-define("IMG_UPLOAD_ORIGINAL","img/upload/origin");
-define("IMG_UPLOAD_MEDIUM","img/upload/medium");
-define("IMG_UPLOAD_SMALL","img/upload/thumb");
+define("IMG_UPLOAD_ORIGINAL","img/upload/origin/");
+define("IMG_UPLOAD_MEDIUM","img/upload/medium/");
+define("IMG_UPLOAD_SMALL","img/upload/thumb/");
+
+// Maximum size for medium images (keep the proportions)
+define("IMG_MEDIUM_WIDTH",950);
+define("IMG_MEDIUM_HEIGHT",700);
+
+// Size for small images (crop the original)
+define("IMG_SMALL_WIDTH",80);
+define("IMG_SMALL_HEIGHT",80);
 
 // Accepted image's formats
 define("IMG_FORMAT",[".jpg", ".jpeg", ".gif", ".png"]);
 
 // Maximum size for image
 define("IMG_MAX_SIZE",100000000); // +- 100 mio
+
+// Qualities for JPG images
+define("IMG_JPG_MEDIUM",90);
+define("IMG_JPG_SMALL",80);
