+++ controller/adminController.php
@@ -2,13 +2,9 @@
 // Dependencies
 require_once "../model/articlesModel.php";
 require_once "../model/usersModel.php";
-require_once "../model/functionDateModel.php";
-require_once "../model/cutTheTextModel.php";
-require_once "../model/paginationModel.php";
-// disconnect
-require_once "../model/disconnectModel.php";
-// images
 require_once "../model/theimagesModel.php";
+
+
 
 // on veut se déconnecter
 if(isset($_GET['p'])&&$_GET['p']=="disconnect"){
@@ -64,13 +60,13 @@ if(isset($_GET['p'])&&$_GET['p']=="create"){
             if($insert){
                 // si on veut y ajouter une image
                 if(!empty($_FILES['theimages_name'])){
-                    $upload = theimagesUpload($_FILES['theimages_name'],IMG_UPLOAD_ORIGINAL,IMG_UPLOAD_MEDIUM,IMG_UPLOAD_SMALL);
+                    $upload = theimagesUpload($_FILES['theimages_name'],IMG_UPLOAD_ORIGINAL,IMG_UPLOAD_MEDIUM,IMG_UPLOAD_SMALL,IMG_MEDIUM_WIDTH,IMG_MEDIUM_HEIGHT,IMG_SMALL_WIDTH,IMG_SMALL_HEIGHT,IMG_JPG_MEDIUM,IMG_JPG_SMALL);
 
                     // l'image a bien été envoyée
                     if(is_array($upload)){
                         // on insert l'image (et on récupère l'id de l'image)
                         $idtheimages = theimagesInsert($db,$_POST['theimages_title'],$upload[0],$insert);
-                        
+
                     }else{
                         $error = $upload;
                     }
