--- model/disconnectModel.php
@@ -1,20 +1,0 @@
-<?php
-// disconnect function
-
-function disconnectModel(){
-    // Destroy session's variables
-    $_SESSION = array();
-
-    // Destroy the cookie
-    if (ini_get("session.use_cookies")) {
-        $params = session_get_cookie_params();
-        setcookie(session_name(), '', time() - 42000,
-            $params["path"], $params["domain"],
-            $params["secure"], $params["httponly"]
-        );
-    }
-
-    // Destroy real session
-    session_destroy();
-}
-
